// src/app/api/model/loginResponse.ts
/**
 * E-Commerce API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';

export interface LoginResponse {
  accessToken: string;
  /**
   * Seconds until expiration
   */
  expiresIn?: number;
  user: User;
}